Extract executor mock into testing package and update tests to use it
- Add internal/runner/executor/testing mock implementation with safe nil
  handling to avoid panics in error-case tests.
- Replace duplicated local MockExecutor implementations with
  executortesting.NewMockExecutor across runner and resource tests.
- Update imports to use executortesting alias and remove redundant mock
  definitions.
- Preserve existing test behavior; pre-commit hooks and go tidy passed.
